Move `WindowContext` to its own file
